// src/app/models/personal.ts
import { Maerespuesta } from "./maerespuesta";
export class Personal {
    codempresa: string;
    codpersonal: string;
    codsucursal: string;
    codtipoplanilla: string;
    numverccostos: string;
    codccostos: string;
    numverareas: string;
    codareas: string;
    empresa: string;
    nomsucursal: string;
    destipoplanilla: string;
    desccostos: string;
    desareas: string;
    nomtrabajador: string;
    apepaterno: string;
    apematerno: string;
    tipdocidentidad: string;
    numdocidentidad: string;
    deslarga: string;
    nhrastrab: string;
    fecingreso: string;
    descargo: string;
    fecnacimiento: string;
    tipsexo: string;
    nacionalidad: string;
    desdireccion: string;
    desdistrito: string;
    desprovincia: string;
    desdepartamento: string;
    numemail: string;
    numtelefono: string;
    horario: string;
    turno: string;
    tipestado: string;
    codcategoria: string;
    codcargo: string;
    condicion: string;
    contratofecini: string;
    antiguedad: string;
    experiencia: string;

    //datos trasient
    edad: string;
    sexo: string;
    isupdate: boolean;
    respuesta: Maerespuesta;

}